+++ src/main/python/plotlyst/view/report/character.py
@@ -17,10 +17,9 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
-from functools import partial
 from typing import Optional, List
 
-from PyQt6.QtCharts import QPieSeries, QPolarChart, QCategoryAxis, QLineSeries, QAreaSeries
+from PyQt6.QtCharts import QPolarChart, QCategoryAxis, QLineSeries, QAreaSeries
 from PyQt6.QtCharts import QValueAxis, QSplineSeries
 from PyQt6.QtCore import pyqtSignal, Qt
 from PyQt6.QtGui import QPen
@@ -30,7 +29,6 @@ from qthandy import clear_layout, vspacer, hbox, transparent, vbox
 
 from src.main.python.plotlyst.core.domain import Novel, Character, Scene, SceneType
 from src.main.python.plotlyst.core.text import html
-from src.main.python.plotlyst.service.cache import acts_registry
 from src.main.python.plotlyst.view.common import icon_to_html_img
 from src.main.python.plotlyst.view.generated.report.character_arc_report_ui import Ui_CharacterArcReport
 from src.main.python.plotlyst.view.generated.report.character_report_ui import Ui_CharacterReport
@@ -45,16 +43,6 @@ class CharacterReport(AbstractReport, Ui_CharacterReport):
 
     def __init__(self, novel: Novel, parent=None):
         super(CharacterReport, self).__init__(novel, parent)
-
-        self.btnAct1.setIcon(IconRegistry.act_one_icon())
-        self.btnAct2.setIcon(IconRegistry.act_two_icon())
-        self.btnAct3.setIcon(IconRegistry.act_three_icon())
-
-        self._povChart = PovDistributionChart()
-        self.chartView.setChart(self._povChart)
-        self.btnAct1.toggled.connect(partial(self._povChart.toggleAct, 1))
-        self.btnAct2.toggled.connect(partial(self._povChart.toggleAct, 2))
-        self.btnAct3.toggled.connect(partial(self._povChart.toggleAct, 3))
 
         self._chartRoles = SupporterRoleChart()
         self.chartViewRoles.setChart(self._chartRoles)
@@ -85,50 +73,11 @@ class CharacterReport(AbstractReport, Ui_CharacterReport):
     @overrides
     def refresh(self):
         self._chartRoles.refresh(self.novel.characters)
-        self._povChart.refresh(self.novel)
         self._chartGenderAll.refresh(self.novel.characters)
         self._chartGenderMajor.refresh(self.novel.major_characters())
         self._chartGenderSecondary.refresh(self.novel.secondary_characters())
         self._chartGenderMinor.refresh(self.novel.minor_characters())
         self._chartAge.refresh(self.novel.characters)
-
-
-class PovDistributionChart(BaseChart):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.createDefaultAxes()
-        self.setTitle("POV Distribution")
-
-        self.pov_number = {}
-        self._acts_filter = {1: True, 2: True, 3: True}
-
-    def toggleAct(self, act: int, toggled: bool):
-        self._acts_filter[act] = toggled
-        self.refresh()
-
-    def refresh(self, novel: Novel):
-        for k in self.pov_number.keys():
-            self.pov_number[k] = 0
-
-        for scene in novel.scenes:
-            if not self._acts_filter[acts_registry.act(scene)]:
-                continue
-            if scene.pov and scene.pov.name not in self.pov_number.keys():
-                self.pov_number[scene.pov.name] = 0
-            if scene.pov:
-                self.pov_number[scene.pov.name] += 1
-
-        series = QPieSeries()
-        for k, v in self.pov_number.items():
-            if v:
-                slice = series.append(k, v)
-                slice.setLabelVisible()
-
-        for slice in series.slices():
-            slice.setLabel(slice.label() + " {:.1f}%".format(100 * slice.percentage()))
-
-        self.removeAllSeries()
-        self.addSeries(series)
 
 
 class AgeChart(PolarBaseChart):
